feat: UnionToTuple
