+++ lib/deep-mark-required/index.ts
@@ -1,14 +1,13 @@
+import { AreNonDistributiveEqual } from "../are-non-distributive-equal";
 import { HasParsablePath } from "../has-parsable-path";
 import { MarkRequired } from "../mark-required";
 import { Paths } from "../paths";
 import { Prettify } from "../prettify";
-import { UnionToIntersection } from "../union-to-intersection";
-
-type AreEqual<First, Second> = First extends Second ? (Second extends First ? true : false) : false;
+import { UnionToTuple } from "../union-to-tuple";
 
 type RecursiveDeepMarkRequiredSinglePath<Type, KeyPath> = KeyPath extends `${infer PropertyKey}.${infer RestKeyPath}`
   ? {
-      [Key in keyof Type]: AreEqual<Key, PropertyKey> extends true
+      [Key in keyof Type]: AreNonDistributiveEqual<Key, PropertyKey> extends true
         ? Prettify<RecursiveDeepMarkRequiredSinglePath<NonNullable<Type[Key]>, RestKeyPath>>
         : Type[Key];
     }
@@ -25,27 +24,6 @@ type RecursiveDeepMarkRequired<Accumulator, KeyPaths extends string[]> = KeyPath
       RecursiveDeepMarkRequired<RecursiveDeepMarkRequiredSinglePath<Accumulator, KeyPath>, RestKeyPaths>
     : never
   : Accumulator;
-
-type LastOfUnion<UnionType> = UnionToIntersection<
-  UnionType extends unknown ? (arg: UnionType) => unknown : never
-> extends (arg: infer LastUnionElement) => unknown
-  ? LastUnionElement
-  : never;
-
-/**
- * A solution to 730. UnionToTuple (hard).
- *
- * Please avoid when possible as it may break at any TypeScript version.
- *
- * @see https://github.com/type-challenges/type-challenges/blob/main/questions/00730-hard-union-to-tuple/README.md
- */
-type UnionToTuple<UnionType, Accumulator extends string[] = []> = [UnionType] extends [never]
-  ? Accumulator
-  : LastOfUnion<UnionType> extends infer LastUnionElement
-  ? LastUnionElement extends string
-    ? UnionToTuple<Exclude<UnionType, LastUnionElement>, [...Accumulator, LastUnionElement]>
-    : never
-  : never;
 
 export type DeepMarkRequired<Type, KeyPathUnion extends Paths<Type>> = HasParsablePath<Type> extends false
   ? Type
